Develop

+++ app/Traits/HasMedia.php
@@ -8,35 +8,38 @@ use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\MorphMany;
 use Illuminate\Http\UploadedFile;
 use Illuminate\Support\Facades\Storage;
-use Livewire\WithFileUploads;
 
 trait HasMedia
 {
-    use WithFileUploads;
 
-    public function setImageAttribute(UploadedFile $file): void
+    public function AddMedia(UploadedFile $file, MediaType $mediaType): Media
     {
-        $this->upload(file: $file, entity: $this, mediaType: MediaType::image);
+        return $this->upload(file: $file, entity: $this, mediaType: $mediaType);
     }
 
-    // set image attribute
-
-    public function upload(UploadedFile $file, Model $entity, MediaType $mediaType)
+    public function upload(UploadedFile $file, Model $entity, MediaType $mediaType): Media
     {
         return $entity->media()->create([
             'mime_type' => $file->getMimeType(),
             'filename' => $file->getClientOriginalName(),
-            'location' => $file->store("{$mediaType->folder()}/{$entity->getTable()}/{$entity->id}", 'public'),
+            'location' => $file->store("{$entity->getTable()}/{$entity->id}/{$mediaType->folder()}", 'public'),
             'custom_property' => $mediaType->value,
             'size' => $file->getSize(),
         ]);
     }
 
-    // upload media
+    // set image attribute
 
     public function media(): MorphMany
     {
         return $this->morphMany(Media::class, 'mediable');
+    }
+
+    // upload media
+
+    public function setImageAttribute(UploadedFile $file): void
+    {
+        $this->upload(file: $file, entity: $this, mediaType: MediaType::image);
     }
 
     public function getImageUrlAttribute(): string
@@ -76,16 +79,20 @@ trait HasMedia
 
     }
 
-    public function delete(): ?bool
+    public function delete(): bool
     {
         // prepare directory
-        $directory = $this->getFirstMedia()->getDirectory();
+        $directory = $this->getFirstMedia()?->getDirectory();
         // delete files
         foreach ($this->media as $media) {
             $media->delete();
         }
         // remove folder and delete model
-        if (Storage::disk('public')->deleteDirectory($directory)) {
+        if ($directory) {
+            if (Storage::disk('public')->deleteDirectory($directory)) {
+                return parent::delete();
+            }
+        } else {
             return parent::delete();
         }
     }
